Break out narrative-report in prep for more properties
