+++ src/components/narrative-report/dto/index.ts
@@ -1,0 +1,1 @@
+export * from './narrative-report.dto';
