+++ src/components/periodic-report/dto/index.ts
@@ -6,3 +6,4 @@ export * from './upload-periodic-report.dto';
 export * from './update-periodic-report.dto';
 export * from './report-period.enum';
 export * from './report-type.enum';
+export * from './merge-periodic-report.dto';
