+++ src/components/periodic-report/dto/periodic-report.dto.ts
@@ -1,6 +1,5 @@
 import { Field, InterfaceType, ObjectType } from '@nestjs/graphql';
 import { keys as keysOf } from 'ts-transformer-keys';
-import { MergeExclusive } from 'type-fest';
 import {
   CalendarDate,
   Resource,
@@ -10,38 +9,25 @@ import {
   SecuredStringNullable,
   Sensitivity,
   SensitivityField,
-  ServerException,
-  simpleSwitch,
 } from '../../../common';
 import { BaseNode as DbBaseNode } from '../../../core/database/results';
 import { ScopedRole } from '../../authorization';
 import { DefinedFile } from '../../file';
 import { ReportType } from './report-type.enum';
 
-type AnyPeriodicReport = MergeExclusive<
-  MergeExclusive<FinancialReport, NarrativeReport>,
-  ProgressReport
->;
-
-export const resolveReportType = (report: Pick<PeriodicReport, 'type'>) => {
-  const type = simpleSwitch(report.type, {
-    Financial: FinancialReport,
-    Narrative: NarrativeReport,
-    Progress: ProgressReport,
-  });
-  if (!type) {
-    throw new ServerException('Could not resolve periodic report type');
-  }
-  return type;
-};
-
 @InterfaceType({
-  resolveType: resolveReportType,
+  resolveType: (obj: PeriodicReport) =>
+    // Prevent circular dependency by lazily importing this.
+    // This file has the concretes which depend on the interface defined here
+    // so this interface file needs to finish loading before the merge file
+    // can be loaded
+    import('./merge-periodic-report.dto').then((m) => m.resolveReportType(obj)),
   implements: [Resource],
 })
 class PeriodicReport extends Resource {
-  static readonly Props = keysOf<PeriodicReport>();
-  static readonly SecuredProps = keysOf<SecuredProps<PeriodicReport>>();
+  static readonly Props: string[] = keysOf<PeriodicReport>();
+  static readonly SecuredProps: string[] =
+    keysOf<SecuredProps<PeriodicReport>>();
 
   @Field(() => ReportType)
   readonly type: ReportType;
@@ -72,10 +58,7 @@ class PeriodicReport extends Resource {
   readonly scope: ScopedRole[];
 }
 
-export {
-  PeriodicReport as IPeriodicReport,
-  AnyPeriodicReport as PeriodicReport,
-};
+export { PeriodicReport as IPeriodicReport };
 
 @ObjectType({
   implements: [PeriodicReport],
@@ -85,16 +68,6 @@ export class FinancialReport extends PeriodicReport {
   static readonly SecuredProps = keysOf<SecuredProps<FinancialReport>>();
 
   readonly type: ReportType.Financial;
-}
-
-@ObjectType({
-  implements: [PeriodicReport],
-})
-export class NarrativeReport extends PeriodicReport {
-  static readonly Props = keysOf<NarrativeReport>();
-  static readonly SecuredProps = keysOf<SecuredProps<NarrativeReport>>();
-
-  readonly type: ReportType.Narrative;
 }
 
 @ObjectType({
